Ajustes UI: Alertas sin tabla histórica, filtro CO2 con decimales, línea roja como promedio en Análisis

+++ src/pages/Historico.jsx
@@ -58,7 +58,7 @@ export default function Historico() {
       if (rowDate > to) return false
     }
     if (searchCo2) {
-      const co2Val = parseInt(searchCo2)
+      const co2Val = parseFloat(searchCo2)
       if (isNaN(co2Val)) return true
       if (row.co2 !== co2Val) return false
     }
